Bruk expect_identical() i staden for expect_equal() om mogleg
expect_equal() should only be used when expect_identical() is
inappropriate, i.e., when x and y need only be numerically equivalent
instead of fully identical (in which case, provide the
tolerance=⁠ argument to expect_equal() explicitly).
This also applies when it's inconvenient to check full equality
(e.g., names can be ignored, in which case ignore_attr = "names"
should be supplied to expect_equal().